add onPhotoLoad callback to thr render method
