+++ packages/react-photo-view/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@dliws/react-photo-view",
-  "version": "1.2.7",
+  "version": "1.2.8",
   "description": "An exquisite React photo preview component",
   "author": "dliws",
   "license": "Apache-2.0",
